# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API g\u00e9rant un cin\u00e9ma",
        "description": "Une API permettant l'acc\u00e8s aux donn\u00e9es de la base via une approche RESTful",
        "termsOfService": "",
        "contact": {
            "name": "Maxime Grimler, Antoine Fidalgo, Florian Laforest, Lionel Bueno",
            "email": "kalimerre@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "myapi.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/film": {
            "get": {
                "tags": [
                    "film"
                ],
                "summary": "Display a listing of films.",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Film"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "film"
                ],
                "summary": "Create a film",
                "description": "Use this method to create a film",
                "operationId": "createFilm",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "titre",
                        "in": "formData",
                        "description": "Name of the film",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "resum",
                        "in": "formData",
                        "description": "Resume of the film",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "date_debut_affiche",
                        "in": "formData",
                        "description": "Date d\u00e9but affiche",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "date_fin_affiche",
                        "in": "formData",
                        "description": "Date fin affiche",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "duree_minutes",
                        "in": "formData",
                        "description": "Dur\u00e9e en minutes",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "annee_production",
                        "in": "formData",
                        "description": "Ann\u00e9e de production",
                        "required": true,
                        "type": "integer",
                        "maximum": "4"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Film created"
                    },
                    "422": {
                        "description": "Champs manquant obligatoire ou incorrect"
                    }
                }
            }
        },
        "/film/{id_film}": {
            "get": {
                "tags": [
                    "film"
                ],
                "summary": "Find film by ID",
                "description": "Returns a single film",
                "operationId": "getFilmById",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_film",
                        "in": "path",
                        "description": "ID of film to return",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "film"
                ],
                "summary": "Delete a film",
                "description": "Delete a film through an ID",
                "operationId": "deleteFilm",
                "parameters": [
                    {
                        "name": "id_film",
                        "in": "path",
                        "description": "Film ID to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film deleted"
                    },
                    "404": {
                        "description": "Invalid film value"
                    }
                }
            }
        }
    },
    "definitions": {
        "Film": {
            "required": [
                "id_film"
            ],
            "properties": {
                "id_film": {
                    "type": "integer"
                },
                "titre": {
                    "type": "string"
                },
                "resum": {
                    "type": "string"
                },
                "date_debut_affiche": {
                    "type": "string",
                    "format": "date"
                },
                "date_fin_affiche": {
                    "type": "string",
                    "format": "date"
                },
                "duree_minutes": {
                    "type": "integer"
                },
                "annee_production": {
                    "type": "integer"
                }
            },
            "xml": {
                "name": "Film"
            }
        }
    }
}